backend/requirements.txt
fastapi
uvicorn[standard]
asyncio-mqtt==0.16.1
paho-mqtt<2.0
sqlmodel
pydantic
pyjwt
python-dotenv
